+++ stroom-meta/stroom-meta-impl-db/src/main/java/stroom/meta/impl/db/MetaDaoImpl.java
@@ -202,8 +202,7 @@ class MetaDaoImpl implements MetaDao, Clearable {
         expressionMapper.map(MetaFields.ID, meta.ID, Long::valueOf);
         expressionMapper.map(MetaFields.META_INTERNAL_PROCESSOR_ID, meta.PROCESSOR_ID, Integer::valueOf);
         expressionMapper.multiMap(MetaFields.FEED, meta.FEED_ID, this::getFeedIds, true);
-        expressionMapper.multiMap(MetaFields.FEED_NAME, meta.FEED_ID, this::getFeedIds);
-        expressionMapper.multiMap(MetaFields.TYPE_NAME, meta.TYPE_ID, this::getTypeIds);
+        expressionMapper.multiMap(MetaFields.TYPE, meta.TYPE_ID, this::getTypeIds);
         expressionMapper.map(MetaFields.PIPELINE, metaProcessor.PIPELINE_UUID, value -> value);
         expressionMapper.map(MetaFields.STATUS,
                 meta.STATUS,
@@ -229,8 +228,7 @@ class MetaDaoImpl implements MetaDao, Clearable {
 
         valueMapper.map(MetaFields.ID, meta.ID, ValLong::create);
         valueMapper.map(MetaFields.FEED, metaFeed.NAME, ValString::create);
-        valueMapper.map(MetaFields.FEED_NAME, metaFeed.NAME, ValString::create);
-        valueMapper.map(MetaFields.TYPE_NAME, metaType.NAME, ValString::create);
+        valueMapper.map(MetaFields.TYPE, metaType.NAME, ValString::create);
         valueMapper.map(MetaFields.PIPELINE, metaProcessor.PIPELINE_UUID, this::getPipelineName);
         valueMapper.map(MetaFields.PARENT_ID, meta.PARENT_ID, ValLong::create);
         valueMapper.map(MetaFields.META_INTERNAL_PROCESSOR_ID, meta.PROCESSOR_ID, ValInteger::create);
@@ -883,16 +881,12 @@ class MetaDaoImpl implements MetaDao, Clearable {
                        final AbstractField[] fields,
                        final Consumer<Val[]> consumer) {
         final List<AbstractField> fieldList = Arrays.asList(fields);
-        final int feedTermCount = ExpressionUtil.termCount(criteria.getExpression(),
-                Set.of(MetaFields.FEED, MetaFields.FEED_NAME));
-        final boolean feedValueExists = fieldList.stream().anyMatch(Set.of(MetaFields.FEED,
-                MetaFields.FEED_NAME)::contains);
-        final int typeTermCount = ExpressionUtil.termCount(criteria.getExpression(), MetaFields.TYPE_NAME);
-        final boolean typeValueExists = fieldList.stream().anyMatch(Predicate.isEqual(MetaFields.TYPE_NAME));
+        final int feedTermCount = ExpressionUtil.termCount(criteria.getExpression(), MetaFields.FEED);
+        final boolean feedValueExists = fieldList.stream().anyMatch(Predicate.isEqual(MetaFields.FEED));
+        final int typeTermCount = ExpressionUtil.termCount(criteria.getExpression(), MetaFields.TYPE);
+        final boolean typeValueExists = fieldList.stream().anyMatch(Predicate.isEqual(MetaFields.TYPE));
         final int processorTermCount = ExpressionUtil.termCount(criteria.getExpression(), MetaFields.PIPELINE);
         final boolean processorValueExists = fieldList.stream().anyMatch(Predicate.isEqual(MetaFields.PIPELINE));
-//        final int extendedTermCount = ExpressionUtil.termCount(
-//        criteria.getExpression(), MetaFields.getExtendedFields());
         final boolean extendedValuesExist = fieldList.stream().anyMatch(MetaFields.getExtendedFields()::contains);
 
         final PageRequest pageRequest = criteria.getPageRequest();
@@ -1329,9 +1323,9 @@ class MetaDaoImpl implements MetaDao, Clearable {
                 field = meta.ID;
             } else if (MetaFields.CREATE_TIME.getName().equals(sort.getId())) {
                 field = meta.CREATE_TIME;
-            } else if (MetaFields.FEED_NAME.getName().equals(sort.getId())) {
+            } else if (MetaFields.FEED.getName().equals(sort.getId())) {
                 field = metaFeed.NAME;
-            } else if (MetaFields.TYPE_NAME.getName().equals(sort.getId())) {
+            } else if (MetaFields.TYPE.getName().equals(sort.getId())) {
                 field = metaType.NAME;
             } else if (MetaFields.PARENT_ID.getName().equals(sort.getId())) {
                 field = meta.PARENT_ID;