+++ stroom-meta/stroom-meta-impl/src/main/java/stroom/meta/impl/MetaServiceImpl.java
@@ -25,6 +25,7 @@ import stroom.meta.shared.Status;
 import stroom.query.api.v2.ExpressionOperator;
 import stroom.query.api.v2.ExpressionOperator.Builder;
 import stroom.query.api.v2.ExpressionTerm;
+import stroom.query.api.v2.ExpressionTerm.Condition;
 import stroom.searchable.api.Searchable;
 import stroom.security.api.SecurityContext;
 import stroom.security.shared.DocumentPermissionNames;
@@ -68,7 +69,6 @@ public class MetaServiceImpl implements MetaService, Searchable {
 
     private final MetaDao metaDao;
     private final MetaFeedDao metaFeedDao;
-    private final MetaTypeDao metaTypeDao;
     private final MetaValueDao metaValueDao;
     private final MetaRetentionTrackerDao metaRetentionTrackerDao;
     private final MetaServiceConfig metaServiceConfig;
@@ -86,7 +86,6 @@ public class MetaServiceImpl implements MetaService, Searchable {
     @Inject
     MetaServiceImpl(final MetaDao metaDao,
                     final MetaFeedDao metaFeedDao,
-                    final MetaTypeDao metaTypeDao,
                     final MetaValueDao metaValueDao,
                     final MetaRetentionTrackerDao metaRetentionTrackerDao,
                     final MetaServiceConfig metaServiceConfig,
@@ -99,7 +98,6 @@ public class MetaServiceImpl implements MetaService, Searchable {
                     final TaskManager taskManager) {
         this.metaDao = metaDao;
         this.metaFeedDao = metaFeedDao;
-        this.metaTypeDao = metaTypeDao;
         this.metaValueDao = metaValueDao;
         this.metaRetentionTrackerDao = metaRetentionTrackerDao;
         this.metaServiceConfig = metaServiceConfig;
@@ -431,8 +429,8 @@ public class MetaServiceImpl implements MetaService, Searchable {
                 .addTerm(MetaFields.EFFECTIVE_TIME,
                         ExpressionTerm.Condition.LESS_THAN,
                         DateUtil.createNormalDateTimeString(criteria.getEffectivePeriod().getToMs()))
-                .addTerm(MetaFields.FEED_NAME, ExpressionTerm.Condition.EQUALS, criteria.getFeed())
-                .addTerm(MetaFields.TYPE_NAME, ExpressionTerm.Condition.EQUALS, criteria.getType())
+                .addTerm(MetaFields.FEED, Condition.EQUALS, criteria.getFeed())
+                .addTerm(MetaFields.TYPE, ExpressionTerm.Condition.EQUALS, criteria.getType())
                 .addTerm(MetaFields.STATUS, ExpressionTerm.Condition.EQUALS, Status.UNLOCKED.getDisplayValue())
                 .build();
 
@@ -451,8 +449,8 @@ public class MetaServiceImpl implements MetaService, Searchable {
                 .addTerm(MetaFields.EFFECTIVE_TIME,
                         ExpressionTerm.Condition.LESS_THAN_OR_EQUAL_TO,
                         DateUtil.createNormalDateTimeString(criteria.getEffectivePeriod().getFromMs()))
-                .addTerm(MetaFields.FEED_NAME, ExpressionTerm.Condition.EQUALS, criteria.getFeed())
-                .addTerm(MetaFields.TYPE_NAME, ExpressionTerm.Condition.EQUALS, criteria.getType())
+                .addTerm(MetaFields.FEED, Condition.EQUALS, criteria.getFeed())
+                .addTerm(MetaFields.TYPE, ExpressionTerm.Condition.EQUALS, criteria.getType())
                 .addTerm(MetaFields.STATUS, ExpressionTerm.Condition.EQUALS, Status.UNLOCKED.getDisplayValue())
                 .build();