+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+* Issue **#2107** : Removed `Feed Name` field and fixed UUID to field name resolution.
+
 * Issue **#2142** : Added debug to help diagnose client cert auth issues.
 
 * Issue **#2107** : Fixed issue where the processor filter UI was saying that no filter had been applied to feeds because the UI wasn't checking feed filtering by docref.